upgrade to rust 31/10/2013
transmute casts on window events whined about wrong sizes,
changed from int to i8, which solves the problem

+++ src/rsfml/graphics/render_window.rs
@@ -391,7 +391,7 @@ impl RenderWindow {
                     0 => false,
                     _ => true
                 };
-                let k : keyboard::Key = unsafe { cast::transmute(self.event.p1 as int) };
+                let k : keyboard::Key = unsafe { cast::transmute(self.event.p1 as i8) };
                 event::KeyPressed{ 
                     code : k, 
                     alt : al, 
@@ -417,7 +417,7 @@ impl RenderWindow {
                     0 => false,
                     _ => true
                 };
-                let k : keyboard::Key = unsafe { cast::transmute(self.event.p1 as int) };
+                let k : keyboard::Key = unsafe { cast::transmute(self.event.p1 as i8) };
                 event::KeyReleased {
                     code : k, 
                     alt : al, 
@@ -434,7 +434,7 @@ impl RenderWindow {
                 }
             },
             8   => {
-                let button : mouse::MouseButton = unsafe {cast::transmute(self.event.p1 as int)};
+                let button : mouse::MouseButton = unsafe {cast::transmute(self.event.p1 as i8)};
                 event::MouseButtonPressed{
                     button : button,
                     x :      unsafe { cast::transmute::<c_uint, c_int>(self.event.p2) as int },
@@ -442,7 +442,7 @@ impl RenderWindow {
                 }
             },
             9   => {
-                let button : mouse::MouseButton = unsafe { cast::transmute(self.event.p1 as int) };
+                let button : mouse::MouseButton = unsafe { cast::transmute(self.event.p1 as i8) };
                 event::MouseButtonReleased{
                     button : button,
                     x :      unsafe { cast::transmute::<c_uint, c_int>(self.event.p2) as int },
@@ -470,7 +470,7 @@ impl RenderWindow {
                 }
             },
             15  => {
-                let ax : joystick::Axis = unsafe { cast::transmute(self.event.p2 as int) };
+                let ax : joystick::Axis = unsafe { cast::transmute(self.event.p2 as i8) };
                 event::JoystickMoved{
                     joystickid : self.event.p1 as uint, 
                     axis : ax,
